Break up type
